refactor responses api helpers
